% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lslx-s3-interface.R
\name{fitted.lslx}
\alias{fitted.lslx}
\title{S3 method to extract model-implied moments from \code{lslx}}
\usage{
\method{fitted}{lslx}(object, selector, lambda, delta, ...)
}
\arguments{
\item{object}{A fitted \code{lslx} object.}

\item{selector}{A \code{character} to specify a selector for determining an optimal penalty level.}

\item{lambda}{A \code{numeric} to specify a chosen optimal lambda value.}

\item{delta}{A \code{numeric} to specify a chosen optimal lambda value.}

\item{...}{Other arguments. For details, please see the \code{$extracted_implied_mean()} and the \code{$extracted_implied_cov()} methods in \code{lslx}.}
}
\description{
\code{fitted.lslx()} is an \code{S3} interface for the \code{$extracted_implied_mean()} and \code{$extracted_implied_cov()} methods to extract model-implied moments from a \code{lslx} object.
}
