% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lslx-s3-interface.R
\name{summary.lslx}
\alias{summary.lslx}
\title{S3 method to summarize \code{lslx} fitting results}
\usage{
\method{summary}{lslx}(object, selector, lambda, delta, ...)
}
\arguments{
\item{object}{A fitted \code{lslx} object.}

\item{selector}{A \code{character} to specify a selector for determining an optimal penalty level.}

\item{lambda}{A \code{numeric} to specify a chosen optimal lambda value.}

\item{delta}{A \code{numeric} to specify a chosen optimal lambda value.}

\item{...}{Other arguments. For details, please see the \code{$summarize()} method in \code{lslx}.}
}
\description{
\code{summary.lslx()} is an \code{S3} interface for summarizing \code{lslx} fitting results.
}
