% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lslx-s3-interface.R
\name{vcov.lslx}
\alias{vcov.lslx}
\title{S3 method to extract covariance matrix of estimates from \code{lslx}}
\usage{
\method{vcov}{lslx}(object, selector, lambda, delta, ...)
}
\arguments{
\item{object}{A fitted \code{lslx} object.}

\item{selector}{A \code{character} to specify a selector for determining an optimal penalty level.}

\item{lambda}{A \code{numeric} to specify a chosen optimal lambda value.}

\item{delta}{A \code{numeric} to specify a chosen optimal lambda value.}

\item{...}{Other arguments. For details, please see the \code{$extracted_coefficient_acov()} method in \code{lslx}.}
}
\description{
\code{vcov.lslx()} is an \code{S3} interface for the \code{$extracted_coefficient_acov()} method to extract covariance matrix of parameter estimate from a \code{lslx} object.
}
