% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelslx-class.R
\docType{class}
\name{prelslx}
\alias{prelslx}
\title{R6 class to obtain preliminary result for semi-confirmatory structural equation modeling}
\value{
Object of \code{prelslx} R6 class.
}
\description{
R6 class to obtain preliminary result for semi-confirmatory structural equation modeling
}
